+++ lib/common_utils.dart
@@ -1,9 +1,10 @@
 library common_utils;
 
+export 'src/data/date_format.dart';
+
 export 'src/object_util.dart';
 export 'src/regex_util.dart';
 export 'src/screen_util.dart';
 export 'src/widget_util.dart';
 export 'src/date_util.dart';
-
-export 'src/data/date_format.dart';
+export 'src/num_util.dart';