+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## 0.0.1
+## 1.0.1
 
-* TODO: CommonUtils initial release.
+* TODO: Newly added NumUtil,Some updates.
+
+## 1.0.0
+
+* TODO: Newly added DateUtil.
 
 ## 0.0.2
 
 * TODO: Contains ScreenUtil, WidgetUtil, ObjectUtil, RegexUtil.
+
+## 0.0.1
+
+* TODO: CommonUtils initial release.